+++ components/KanjinizedDate.vue
